feat: create SuspiciousUsers module (Phase 3.3)\n\n- Move SuspiciousUsersStorage from Services to SuspiciousUsers namespace\n- Create SuspiciousUsersModule with DI registration\n- Update Program.cs to use new module\n- Fix using directives in test files\n- All tests pass (1005/1010, 5 skipped due to API keys)\n\nProgress: 18/18 modules created, Program.cs reduced from 469 to ~300 lines
