+++ ClubDoorman/Services/SuspiciousUsers/SuspiciousUsersModule.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace ClubDoorman.Services.SuspiciousUsers;
+
+public static class SuspiciousUsersModule
+{
+    public static IServiceCollection AddSuspiciousUsersServices(this IServiceCollection services)
+    {
+        // TODO: Добавить регистрацию сервисов если необходимо
+        return services;
+    }
+}
